refactor(types): add keycode and modifier mod
